+++ cns/podprovider/statefile_test.go
@@ -1,4 +1,4 @@
-package cnireconciler
+package podprovider
 
 import (
 	"os"
